backend/requirements.txt
asgiref
Django
django-cors-headers
gunicorn
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
djangorestframework
djangorestframework-simplejwt